move single entry form to shared component folder
