+++ src/app/(default)/editClimb/[slug]/components/ClimbLocationForm.tsx
@@ -1,7 +1,7 @@
 'use client'
 import { useSession } from 'next-auth/react'
 
-import { SingleEntryForm } from '@/app/(default)/editArea/[slug]/components/SingleEntryForm'
+import { SingleEntryForm } from '@/app/(default)/components/AreaAndClimb/SingleEntryForm'
 import { CLIMB_LOCATION_FORM_VALIDATION_RULES } from '../validation'
 import useUpdateClimbsCmd from '@/js/hooks/useUpdateClimbsCmd'
 import { MarkdownTextArea } from '@/components/ui/form/MarkdownTextArea'
