+++ src/app/(default)/editClimb/[slug]/components/SidebarNav.tsx
@@ -6,7 +6,7 @@ import { TotalLengthInput } from '@/components/edit/form/TotalLengthInput'
 import { LegacyFAInput } from '@/components/edit/form/LegacyFAInput'
 import Grade from '@/js/grades/Grade'
 import useUpdateClimbsCmd from '@/js/hooks/useUpdateClimbsCmd'
-import { SubmitButton } from '../../../editArea/[slug]/components/SingleEntryForm'
+import { SubmitButton } from '../../../components/AreaAndClimb/SingleEntryForm'
 import { ArticleLastUpdate } from '@/components/edit/ArticleLastUpdate'
 import { removeTypenameFromDisciplines } from '@/js/utils'
 import { ClimbDisciplineRecord, ClimbType } from '@/js/types'
